AtCoder/ABC341/D/main_test.go
package main

import (
	"bytes"
	"io"
	"strings"
	"testing"
)

func Test_run(t *testing.T) {
	type args struct {
		stdin io.Reader
	}
	tests := []struct {
		name    string
		args    args
		wantOut string
	}{
		{name: "1", args: args{stdin: strings.NewReader(`2 3 5`)}, wantOut: `9`},
		{name: "2", args: args{stdin: strings.NewReader(`1 2 3`)}, wantOut: `5`},
		{name: "3", args: args{stdin: strings.NewReader(`100000000 99999999 10000000000`)}, wantOut: `500000002500000000`},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			out := &bytes.Buffer{}
			run(tt.args.stdin, out)
			if gotOut := out.String(); gotOut != tt.wantOut {
				t.Errorf("run() = %v, want %v", gotOut, tt.wantOut)
			}
		})
	}
}
